+++ django_db_sanitizer/fetchers/base.py
@@ -41,6 +41,9 @@ class BaseFetcher(object):
         qs = self.get_queryset_manager()\
             .filter(**self.filters_for_fetching)\
             .exclude(**self.excludes_for_fetching)
+        if qs.count() == 0:
+            logger.warning("{0} is showing 0 results for its fetching query! "
+                           "This is the query : '{1}'".format(self, qs.query))
         return qs
 
     def get_filtered_queryset_values(self, field_names, include_pk_field=True):
@@ -57,8 +60,9 @@ class BaseFetcher(object):
 
         if include_pk_field:
             pk_field = self.get_model_pk_field_name()
-            required_fields = (pk_field, *tuple(f for f in field_names
-                                                if f != pk_field))
+
+            required_fields = (pk_field,) + tuple(f for f in field_names
+                                                  if f != pk_field)
         else:
             required_fields = tuple(f for f in field_names)