+++ test_project/test_app/models.py
@@ -5,13 +5,21 @@ from django.utils.six import python_2_unicode_compatible
 from test_app.constants import MONTHS, DAY_OF_BIRTH
 
 
-@python_2_unicode_compatible
-class Profile(models.Model):
-    user = models.OneToOneField(User, related_name="profile")
-    card_number = models.CharField(max_length=250, null=True, blank=True)
+class AbstractProfile(models.Model):
 
-    phone = models.CharField(max_length=100, blank=True, default="")
-    mobile_phone = models.CharField(max_length=100, blank=True, default="")
+    class Meta:
+        abstract = True
+
+
+@python_2_unicode_compatible
+class Profile(AbstractProfile):
+    user = models.OneToOneField(User, related_name="profile")
+    card_number = models.CharField(max_length=250, null=True, blank=True,
+                                   default="")
+    initials = models.CharField(max_length=4, blank=True, default="")
+
+    phone = models.CharField(max_length=100, blank=False, default="911")
+    mobile_phone = models.CharField(max_length=100, blank=False, unique=True)
 
     address = models.CharField(max_length=500, blank=True, default="")
     city = models.CharField(max_length=250, blank=True, default="")
@@ -19,23 +27,25 @@ class Profile(models.Model):
     state = models.CharField(max_length=100, blank=True, default="")
     country = models.CharField(max_length=100, blank=True, default="")
 
-    awesomeness_rank = models.IntegerField(blank=True, default=0)
+    awesomeness_rank = models.BigIntegerField(blank=True, default=0)
     importance_rank = models.IntegerField(blank=True, default=0)
 
-    number_of_cars = models.IntegerField(blank=True, default=0)
-    number_of_computers = models.IntegerField(blank=True, default=0)
+    number_of_cars = models.PositiveIntegerField(blank=True, default=0)
+    number_of_computers = models.PositiveSmallIntegerField(blank=True,
+                                                           default=0)
 
     month_of_birth = models.CharField(
         max_length=3, choices=MONTHS, blank=True, null=True)
     day_of_birth = models.IntegerField(
         choices=DAY_OF_BIRTH, blank=True, null=True)
-    year_of_birth = models.IntegerField(blank=True, null=True)
+    year_of_birth = models.SmallIntegerField(blank=True, null=False)
     date_of_birth = models.DateField(blank=True, null=True)
 
     money_balance = models.DecimalField(
         max_digits=14, decimal_places=2, default=0)
 
     internal_notes = models.TextField(default="", blank=True)
+    admin_notes = models.TextField(default="", blank=True, max_length=500)
 
     is_subscribed_to_mailing = models.BooleanField(default=True)