+++ CHANGELOG.rst
@@ -3,7 +3,11 @@ Changelog - django-db-sanitizer
 ===============================
 
 
-0.1.0 - TBD
------------
+0.1.0 - April 27th 2016
+-----------------------
 
 - First release!
+- Can configure Django apps to sanitize/anonymize data fron specific fields
+  of specific models.
+- Added a number of basic Sanitizer classes covering some basic sanitizing
+  needs for basic Django ModelField types.